refactor: move the Examples to folder PIDControl
