--- Examples/fiware_groups.json
@@ -1,22 +1,0 @@
-[
-  {
-    "service": "controller",
-    "subservice": "/",
-    "resource": "/iot/json",
-    "apikey": "buildings",
-    "timestamp": null,
-    "entity_type": null,
-    "trust": null,
-    "cbHost": null,
-    "lazy": [],
-    "commands": [],
-    "attributes": [],
-    "static_attributes": [],
-    "internal_attributes": [],
-    "expressionLanguage": "legacy",
-    "explicitAttrs": false,
-    "autoprovision": true,
-    "ngsiVersion": "v2",
-    "defaultEntityNameConjunction": null
-  }
-]
